refactor(variance): separate shared alert helpers
The variance facade still owns calculation and alert management, but shared alert data normalization and small query/config helpers no longer need to live inline with the service classes. Move them into a sibling helper module while keeping the facade behavior unchanged.

Constraint: Keep AlertManagementService in the facade until the next extraction commit

Rejected: Extract AlertManagementService in the same commit | a helper-only commit keeps the review surface small and independently verified

Confidence: high

Scope-risk: narrow

Tested: npx vitest run --config vitest.config.mjs --configLoader native --project=server tests/unit/services/variance-tracking.test.ts

Tested: npm run check

Not-tested: Full npm test suite
